+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Python version `3.11` tested during CI
+- Python version `3.11` added to package classifiers
+
 ### Removed
 
 - Support for Python version `3.7`.