+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Python version `3.11` tested during CI
 - Python version `3.11` added to package classifiers
+- Python version `3.12` tested during CI
+- Python version `3.12` added to package classifiers
 
 ### Removed